Improve objects communication
