+++ task_2_3_1/src/main/java/kozlov/kirill/snake/model/game/snake/ComputerSnakeAi.java
@@ -1,5 +1,6 @@
-package kozlov.kirill.snake.model.game;
+package kozlov.kirill.snake.model.game.snake;
 
+import kozlov.kirill.snake.model.game.Vector;
 import lombok.Getter;
 
 public abstract class ComputerSnakeAi {
