+++ task_2_3_1/src/main/java/kozlov/kirill/snake/model/game/snake/Snake.java
@@ -1,8 +1,12 @@
-package kozlov.kirill.snake.model.game;
+package kozlov.kirill.snake.model.game.snake;
 
 import java.util.LinkedList;
 import java.util.List;
 
+import kozlov.kirill.snake.model.game.Field;
+import kozlov.kirill.snake.model.game.FieldObject;
+import kozlov.kirill.snake.model.game.Point;
+import kozlov.kirill.snake.model.game.Vector;
 import lombok.Getter;
 
 /**
